fix test layout
